
\name{TwoPhaseScreenTest-package}
\alias{TwoPhaseScreenTest}
\docType{package}
\title{Two-Phase Designs for Cost-Effective Evaluation of Cancer Screening Tests}
\description{A subsampling strategy that strategically selects biospecimens
to evaluate a new cancer screening test. The goal of the sampling strategy is to optimize statistical
efficiency in estimating the association between test results and prevalent/incident outcomes, as well as
the cumulative risks associated with each test strata.}
\details{
  The main functions are \code{\link{twoPhaseScreen_design}}, \code{\link{twoPhaseScreen_est}}, 
  and \code{\link{twoPhaseScreen_predict}}.
}
\author{Fangya Mao <fangya.mao@nih.gov>, Richard Cook <rjcook@uwaterloo.ca>, Li Cheung <li.cheung@nih.gov>}
\keyword{package}
\references{
  Mao, F., Cook R. J., Lorey T., Wentzensen N., Cheung, L.C. (2024+).
  Two-Phase Designs for Cost-Effective Evaluation of Cancer Screening Tests.
}

